+++ docs/src/history.md
@@ -6,6 +6,7 @@
 * ADDED: `>>` and `<<` methods for `ModUInt`.
 * ADDED: `% type` for truncating `MLUInt` to built-in types.
 * ADDED: `mulmod_montgomery_ct()` for constant-time Montgomery multiplication.
+* ADDED: `rand()` for `MLUInt` and its ranges.
 
 
 ## v0.1.2 (19 February 2020)